requirements/runtime.txt
absl-py
accelerate>=0.19.0
cpm_kernels
datasets>=2.12.0,<=3.6.0
einops>=0.5.0
evaluate>=0.3.0
func_timeout
fuzzywuzzy
gradio-client
h5py
httpx
immutabledict
importlib-metadata
jieba
json5
jsonlines
mmengine-lite
nltk>=3.7
numpy>=1.23.4,<2.0.0
openai
opencv-python-headless
orjson
pandas<2.0.0
plotly
prettytable
protobuf
pyext==0.5
python-Levenshtein
rank_bm25==0.2.2
rapidfuzz
requests>=2.31.0
retrying
rich
rouge
rouge_chinese
rouge_score
sacrebleu
scikit_learn==1.5.0
scipy
seaborn
tabulate
tiktoken
timeout_decorator
tokenizers
torch
tqdm>=4.64.1
transformers
tree-sitter==0.21.3
tree_sitter_languages>=1.10.2
typer
Pillow==11.2.1
